VI/requirements.txt
altair
pandas
jupyter
vega_dataset
streamlit